pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tracer"
version = "0.1.0"
description = "GPU-aware trace generator for LLM inference"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.32,<3",
  "python-dotenv>=1.0,<2"
]

[project.scripts]
tracer = "tracer.tracer:main"

[tool.setuptools]
packages = ["tracer"]
